Inventory.py with bug fixes
Bug fixed with option 3 - accidentally called the wrong function.
Bug fixed when using option 3 followed by any other function - restocking an item left a newline character on the end of the added string in the file, creating a line with length 0. This would cause a 'line index out of range' error when attempting to loop through that line.